Fix issue with card::fieldid_r being saved in uint16 global variables
This caused issues in the various checks for an attack validity, as with the fieldid being uin32, it was being truncated when stored in the global state variables, and thus when it was compared again to them, the check would never return true as the uint16 was beign promoted to uint32 and thus having it impossible to match the other value

+++ card.h
@@ -75,7 +75,7 @@ public:
 	using relation_map = std::unordered_map<card*, uint32_t>;
 	using counter_map = std::map<uint16_t, std::array<uint16_t, 2>>;
 	using effect_count = std::map<uint32_t, int32_t>;
-	class attacker_map : public std::unordered_map<uint16_t, std::pair<card*, uint32_t>> {
+	class attacker_map : public std::unordered_map<uint32_t, std::pair<card*, uint32_t>> {
 	public:
 		void addcard(card* pcard);
 		uint32_t findcard(card* pcard);
